mgmt: Logging abandon volumes/snapshots
Logging StorPool's volumes, snapshots and snapshots for a templates which are left on StorPool and CloudStack does
not know about them. Logs for the left objects are kept in "/var/log/cloudstack/management/storpool-abandoned-objects.log" on managment server

+++ src/org/apache/cloudstack/storage/helper/StorPoolDeleteTemplateCmd.java
@@ -106,11 +106,15 @@ public class StorPoolDeleteTemplateCmd extends BaseAsyncCmd {
                 SpConnectionDesc conn = new SpConnectionDesc(obj.getDataStore().getUuid());
                 if (StorpoolUtil.snapshotExists(StorpoolStorageAdaptor.getVolumeNameFromPath(template.getLocalDownloadPath(), true), conn)) {
                     SpApiResponse resp = StorpoolUtil.snapshotDelete(StorpoolStorageAdaptor.getVolumeNameFromPath(template.getLocalDownloadPath(), true), conn);
-                    VMTemplateDetailVO detail = vmTemplateDetailDao.findDetail(template.getTemplateId(), StorpoolUtil.SP_STORAGE_POOL_ID);
-                    if (detail != null) {
-                        vmTemplateDetailDao.remove(detail.getId());
+                    if (resp.getError() == null) {
+                        VMTemplateDetailVO detail = vmTemplateDetailDao.findDetail(template.getTemplateId(), StorpoolUtil.SP_STORAGE_POOL_ID);
+                        if (detail != null) {
+                            vmTemplateDetailDao.remove(detail.getId());
+                        }
+                        StorpoolUtil.spLog("Deleted template from StorPool %s", template.getLocalDownloadPath());
+                    }else {
+                        StorpoolUtil.spLog("Could not delete template from StorPool %s due to %s", template.getLocalDownloadPath(), resp.getError());
                     }
-                    StorpoolUtil.spLog("Delete template from StorPool. Snapshot name=%s, result=%s", obj.getUuid(), resp.getError());
                 }
             }
         }catch (CloudRuntimeException e) {
